8265473: Move os::Linux to its own header file
